Gather update exts feature into exts_tools.py file
