+++ easybuild/framework/easyconfig/tools.py
@@ -41,10 +41,8 @@ Authors:
 import copy
 import fnmatch
 import glob
-import hashlib
 import os
 import re
-import requests
 import sys
 import tempfile
 from collections import OrderedDict
@@ -71,19 +69,6 @@ from easybuild.tools.toolchain.toolchain import is_system_toolchain
 from easybuild.tools.toolchain.utilities import search_toolchain
 from easybuild.tools.utilities import only_if_module_is_available, quote_str
 from easybuild.tools.version import VERSION as EASYBUILD_VERSION
-
-# URLs for package repositories
-CRANDB_URL = "https://crandb.r-pkg.org"
-CRANDB_CONTRIB_URL = "https://cran.r-project.org/src/contrib"
-PYPI_URL = "https://pypi.org/pypi"
-BIOCONDUCTOR_URL = "https://bioconductor.org/packages"
-BIOCONDUCTOR_PKGS_URL = "bioc/packages.json"
-BIOCONDUCTOR_ANNOTATION_URL = "data/annotation/packages.json"
-BIOCONDUCTOR_EXPERIMENT_URL = "data/experiment/packages.json"
-
-# Global variable to store Bioconductor packages
-bioc_packages_cache = None
-
 
 # optional Python packages, these might be missing
 # failing imports are just ignored
@@ -916,297 +901,3 @@ def det_copy_ec_specs(orig_paths, from_pr=None, from_commit=None):
                 raise EasyBuildError("Found multiple paths for %s in commit: %s", filename, commit_matches)
 
     return paths, target_path
-
-
-def calculate_md5(data):
-    """
-    Calculate the MD5 checksum of the given data
-
-    :param data: data to calculate the checksum for
-
-    :return: MD5 checksum of the given data
-    """
-
-    if not data:
-        raise EasyBuildError("No data provided to calculate the MD5 checksum")
-
-    # calculate the md5 checksum
-    md5_hash = hashlib.md5()
-    md5_hash.update(data)
-
-    return md5_hash.hexdigest()
-
-
-def get_R_pkg_checksum(pkg_metadata, bioconductor_version=None):
-    """
-    Get the checksum of the given R package version
-
-    :param pkg_metadata: package metadata
-    :param bioconductor_version: bioconductor version to use (if any)
-
-    :return: checksum of the given R package version
-    """
-
-    if not pkg_metadata:
-        raise EasyBuildError("No R package metadata provided to get the checksum")
-
-    # check if checksum is provided in the metadata
-    checksum = pkg_metadata.get('MD5sum', '')
-
-    # if checksum is not provided, calculate it
-    if not checksum:
-        # get package information
-        pkg_name = pkg_metadata.get('Package', '')
-        pkg_version = pkg_metadata.get('Version', '')
-
-        # download package from the database
-        package = get_pkg('RPackage', pkg_name, pkg_version, bioconductor_version)
-
-        # calculate the checksum
-        if package:
-            checksum = calculate_md5(data=package)
-        else:
-            print_warning("Failed to download package %s v%s to calculate the checksum" % (pkg_name, pkg_version), log=_log)
-            checksum = ''
-
-    return checksum
-
-
-def get_python_pkg_checksum(pkg_metadata):
-    """
-    Get the checksum of the given Python package version
-
-    :param pkg_metadata: package metadata
-
-    :return: checksum of the given Python package version
-    """
-
-    if not pkg_metadata:
-        raise EasyBuildError("No python package metadata provided to get the checksum")
-
-    # initialize variable
-    checksum = ''
-
-    # get the data of the given package version
-    pkg_name = pkg_metadata.get('info', {}).get('name', '')
-    pkg_version = pkg_metadata.get('info', {}).get('version', '')
-    pkg_releases = pkg_metadata.get('releases', {})
-    pkg_info = pkg_releases.get(pkg_version, [])
-
-    # parse the version info to get the checksum
-    if pkg_info:
-        # look for sdist first
-        for file_info in pkg_info:
-            if file_info.get('packagetype') == 'sdist':
-                checksum = file_info.get('digests', {}).get('sha256', '')
-
-        # if no sdist found, take the checksum of the first distribution file
-        if not checksum:
-            checksum = pkg_info[0].get('digests', {}).get('sha256', '')
-
-        if not checksum:
-            print_warning("Failed to get package %s v%s checksum from distribution file" %
-                          (pkg_name, pkg_version), log=_log)  
-    else:
-        print_warning("Failed to get package %s v%s information to grasp the checksum" %
-                                (pkg_name, pkg_version), log=_log)
-
-    return checksum
-
-
-def get_bioconductor_pkgs_metadata(bioc_version):
-    """
-    Get the list of Bioconductor packages from the Bioconductor database.
-
-    :param bioc_version: Bioconductor version
-
-    :return: list of Bioconductor packages
-    """
-
-    if not bioc_version:
-        raise EasyBuildError("No bioconductor version provided to get the bioconductor packages")
-
-    # global variable to store the bioconductor packages
-    global bioc_packages_cache
-
-    # bioconductor URLs
-    bioc_urls = ['%s/json/%s/%s' % (BIOCONDUCTOR_URL, bioc_version, BIOCONDUCTOR_PKGS_URL),
-                 '%s/json/%s/%s' % (BIOCONDUCTOR_URL, bioc_version, BIOCONDUCTOR_ANNOTATION_URL),
-                 '%s/json/%s/%s' % (BIOCONDUCTOR_URL, bioc_version, BIOCONDUCTOR_EXPERIMENT_URL)]
-
-    # check if the packages are already stored in memory
-    if bioc_packages_cache is None:
-
-        # initialize the cache
-        bioc_packages_cache = {}
-
-        # retrieve packages from the cloud
-        for url in bioc_urls:
-            try:
-                response = requests.get(url)
-
-                if response.status_code == 200:
-                    bioc_packages_cache.update(response.json())
-                else:
-                    print_warning(
-                        f"Failed to get biocondcutor packages from {url}: HTTP status: {response.status_code}")
-            except Exception as err:
-                print_warning(f"Exception while getting bioconductor packages from  {url}: {err}")
-
-    return bioc_packages_cache
-
-
-def get_pkg(pkg_class, pkg_name, pkg_version, bioconductor_version=None):
-    """
-    Get the package from the database
-
-    : param pkg_class: package class (RPackage)
-    : param pkg_name: package name
-    : param pkg_version: package version
-    : param bioconductor_version: bioconductor version (if any)
-
-    : return: package from database
-    """
-
-    if not pkg_class:
-        raise EasyBuildError("No package class provided to get the package")
-    
-    if not pkg_name:
-        raise EasyBuildError("No package name provided to get the package")
-    
-    if not pkg_version:
-        raise EasyBuildError("No package version provided to get the package")
-
-    # initialize variables
-    urls = []
-    pkg = None
-
-    # build the url to get the package from the database
-    if pkg_class == "RPackage":
-        # URL for CRANDB
-        crandb_url = "%s/%s_%s.tar.gz" % (CRANDB_CONTRIB_URL, pkg_name, pkg_version)
-        urls.append(crandb_url)
-
-        # URL for CRANDB Archive
-        crandb_archive_url = "%s/Archive/%s/%s_%s.tar.gz" % (CRANDB_CONTRIB_URL, pkg_name, pkg_name, pkg_version)
-        urls.append(crandb_archive_url)
-
-        # URL for Bioconductor package
-        if bioconductor_version:
-            # Construct the URL for Bioconductor package
-            url = "%s/%s/bioc/src/contrib/%s_%s.tar.gz" % (BIOCONDUCTOR_URL,
-                                                           bioconductor_version, pkg_name, pkg_version)
-            urls.append(url)
-    else:
-        raise EasyBuildError("get_pkg function only supports RPackage extensions")
-
-    try:
-        for url in urls:
-            # get the package's metadata from the database
-            response = requests.get(url, stream=True)
-
-            if response.status_code == 200:
-                pkg = response.content
-                break
-
-    except Exception as err:
-        print_warning("Exception while downloading package %s v%s. Error: %s" % (pkg_name, pkg_version, err))
-
-    return pkg
-
-
-def get_pkg_metadata(pkg_class, pkg_name, pkg_version=None, bioc_version=None):
-    """
-    Get the metadata of the given package
-
-    :param pkg_class: package class (RPackage, PythonPackage)
-    :param pkg_name: package name
-    :param pkg_version: package version. If None, the latest version will be retrieved.
-    :param bioc_version: bioconductor version (if any)
-
-    :return: package metadata
-    """
-
-    # initialize variables
-    pkg_metadata = None
-    bioc_packages = None
-
-    if not pkg_name:
-        raise EasyBuildError("No package name provided to get the package metadata")
-
-    # build the url to get the metadata from the database
-    if pkg_class == "RPackage":
-        if pkg_version:
-            url = "%s/%s/%s" % (CRANDB_URL, pkg_name, pkg_version)
-        else:
-            url = "%s/%s" % (CRANDB_URL, pkg_name)
-
-        # get bioc packages if bioconductor version is provided
-        if bioc_version:
-            bioc_packages = get_bioconductor_pkgs_metadata(bioc_version)
-
-    elif pkg_class == "PythonPackage":
-        url = "%s/%s/json" % (PYPI_URL, pkg_name)
-
-    else:
-        raise EasyBuildError("exts_defaultclass %s not supported" % pkg_class)
-
-    try:
-        # get the package's metadata from the database
-        response = requests.get(url)
-
-        if response.status_code == 200:
-            pkg_metadata = response.json()
-
-    except Exception as err:
-        print_warning("Exception while getting metadata for extension %s: %s" % (pkg_name, err))
-
-    # if the package is not found in the database, then check if it is a bioconductor package
-    if not pkg_metadata and bioc_packages:
-        # iterate over bioconductor packages to find the package
-        for package in bioc_packages.items():
-            if package[0] == pkg_name:
-                pkg_metadata = package[1]
-                break
-
-    return pkg_metadata
-
-
-def format_metadata_as_extension(pkg_class, pkg_metadata, bioconductor_version=None):
-    """
-    Get the package metadata as an exts_list extension format
-
-    :param pkg_class: package class (RPackage, PythonPackage, PerlPackage)
-    :param pkg_metadata: package metadata
-    :param bioconductor_version: bioconductor version
-
-    :return: package metadata in exts_list extension format
-    """
-
-    if not pkg_metadata:
-        raise EasyBuildError("No package metadata provided to format as extension")
-
-    # check the package class and parse the metadata accordingly
-    if pkg_class == "RPackage":
-        name = pkg_metadata.get('Package', '')
-        version = pkg_metadata.get('Version', '')
-        checksum = get_R_pkg_checksum(pkg_metadata, bioconductor_version)
-
-    elif pkg_class == "PythonPackage":
-        name = pkg_metadata.get('info', {}).get('name', '')
-        version = pkg_metadata.get('info', {}).get('version', '')
-        checksum = get_python_pkg_checksum(pkg_metadata)
-
-    else:
-        raise EasyBuildError("exts_defaultclass %s not supported" % pkg_class)
-
-    # remove any non-alphanumeric characters from the version
-    allowed_version_chars = r'[^0-9><=!*. \-]'
-    version = re.sub(allowed_version_chars, '', version)
-
-    # remove any new line characters
-    name = name.replace('\n', '')
-    version = version.replace('\n', '')
-    checksum = checksum.replace('\n', '')
-
-    return {"name": name, "version": version,  "options": {"checksums": [checksum]}}
